remoteproc: account for virt addresses in sh_addr
Currently we hope to find a physicall address there, however it's not
always the case - elf headers contain virtual addresses there.

So, to fix that, we calculate the phys addr by getting the offset between
the first segment and our segment, and add it to the entry point.

That's extremily shitty, but I currently don't know any better, and, it
works.

Signed-off-by: Veaceslav Falico <vfalico@gmail.com>