dummy_proc: add an BSP callback setter and confirm booting of AP
So, currently we rely only on a delay, and we don't know for sure if the AP
booted or not.

To fix that - make it possible to set the BSP IPI callback, and use it as a
notification mechanism of AP booting (AP will always send one IPI in the
late kernel booting stage). Also, add a parameter on how long to wait for
the confirmation from AP. Finally, set a generic callback for BSP IPI
(which will be, most probably, used for virtio).

The logic now is the following:

	BSP				AP
---------------------		---------------------
Set up mem/etc.
Set IPI callback to boot
Start the CPU			Start booting...
Wait for boot callback...
				Boot...
				End: send IPI to BSP
... got callback
Set generic callback
Set up vrings
Send IPI to AP
				Got IPI: vrings are ready
				set up vrings [TBD]

Afterwards, we are sure that AP has booted, and we are sure at which moment
we're ready to use the vrings on AP.

Signed-off-by: Veaceslav Falico <vfalico@gmail.com>