apic: remove lapic hack
Currently, to overcome the situation where AP reconfigure the BSP's APIC
(and, thus, breaks irq routing and other stuff for BSP) - popcorn uses the
lapic_is_bsp() and friends to find out who is the AP/BSP and, in case of
BSP, forbid changing the APIC.

Remove lapic_is_bsp() and boot with noapic currently, as anyway it's
broken.

This issue must be fixed somehow, as we need to get at least some IRQs
mapped for different devices, but currently I don't see a way to do it - as
there's no way to "forbid" overwriting already-existing settings, and
inventing yet another hacks in that deep code isn't a good idea at all.

Signed-off-by: Veaceslav Falico <vfalico@gmail.com>