dummy_lproc: populate cpu present/online/active masks
This commit, basically, gets rid of the core modifications related to CPU
filtering in setup_arch() and other early boots.

We limit the masks by overloading the... x86_init.oem.banner(), which
is the closest arch-related function which we can modify after the cpu
masks have been inited.

This hack is needed because kernel has only two ways of booting on any CPU:

1) BSP, where the kernel populates all the structures "use everything we've
   got".
2) APs, where the kernel has already all the needed structures populated
   and initializes only the CPU it's booted on.

And there is no middle ground, where we'd initilize everything except
what's banned.

Signed-off-by: Veaceslav Falico <vfalico@gmail.com>